update general styling

+++ client/src/Helpers/formatDate.js
@@ -2,7 +2,7 @@
 export default function formatDate(dateTimeString){
 
     const dateObj = new Date(dateTimeString)
-    const month = dateObj.toLocaleString('default', {month: 'long'})
+    const month = dateObj.toLocaleString('default', {month: 'short'})
     const day = dateObj.getDate();
     const year = dateObj.getFullYear()
     return `${month} ${day}, ${year}`
